lambda/api_handler/requirements.txt
boto3==1.38.18
PyJWT==2.10.1 